Add warnings for bad configurations
Warnings are now logged when key configurations are not set. We require
that a lot of these are set in production apps, and it's good to remind
developers if they're missing something.

I've also noted that the old n-express environment variables are
deprecated, we'll be removing them in the next major version.

See #40